requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
arq==0.25.0
redis==5.0.4
reportlab==4.1.0
pydantic==2.7.1
pydantic-settings==2.2.1
dnspython==2.6.1
